Test that buildxyceplugin defines __XYCE__
In issue Xyce#158, we had a request that buildxyceplugin quietly add a
define of a unique symbol that one could use in a Verilog-A model to
stick in Xyce-specific hacks.

It is unfortunate that such hacks are necessary, but sometimes they
are and it behooves us to throw a bone to model developers who might
want to do so.

In commit Xyce@eace36bc8 I added "-D__XYCE__" to the admsXml command
invoked by buildxyceplugin.

In this commit, I test it.

In my previous commit, rlc2.va was used in a simple netlist to produce
a bunch of output that was captured as gold standard.

In this commit, I add an "ifdef __XYCE__" and an else block such that
if the symbol is NOT defined the answers will be dramatically wrong
because the resistor and inductor currents will have been multiplied
by five.  This will lead to a compare failure.

This test is only run in shared library builds, which are only done
nightly.  But it should at least catch any regressions.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/158
Xyce#158