Loosen absolute tolerance on rlc_series test
My gold standard for the VERILOGPLUGIN/rlc_series test was generated
on BSD, and when compared with default tolerances to the output on
Linux it is getting slightly too much relative error in some signals
that are rather small to begin with.  I am loosening the absolute
tolerance on this (which is there just to prevent divisions by zero
when the gold standard is close to zero) to force the relative error
to be a little less sensitive.

Xyce#158
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/158