Excluding these as now .param supports specials
I recently updated the source code so that .param can support special
variables such as TEMP, VT, TIME, FREQ, etc.

This is related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/157

These two tests were to ensure that if someone attempted to set a .param
to depend on a special variable that Xyce would emit an appropriate
error message.  As this is now allowed, these tests would fail.  So,
excluding.