Add a test for max time step out of a break point
The test is used to check if the first step out of a break point
is limited by the maximum step specified in the netlist. The
max step was ignored previously and this if fixed. The test failed
before the issue is fixed. It now passes after the fixes for #159