Adding Basker tests for both scalar and block solver
The templated Basker solver in Amesos2 can provide a direct solver
for block analysis types in two ways.  First, the scalar implementation
of the solver can be used to perform time or frequency-domain solves.
Second, the scalar type can be specialized such that each entry in the
matrix is a block.  Then the solver can provide a block factorization
for time or frequency-domain solves.

This commit adds tests, one for each Basker and block Basker solver,
that are guarded by a new amesos2basker tag that is required for these
tests.