Add detection of Basker to suggestXyceTagList.sh
@hkthorn has added tests that require an "amesos2basker" tag to run.

This commit adds a scan of the Xyce -capabilities output to detect
when a binary has this compiled in, and to enable that tag if so.

A similar commit was made to Jenkins scripts in
JenkinsScripts@65bd2c6.  Most of those don't use suggestXyceTagList.sh
yet, but rather their own internal script snippets that were used
to *create* suggestXyceTagList.sh.  This script *is* used in TDEV and
Shared testing, though.

This commit assures that those builds will run the new test.

There does not appear to be an issue associated with the tests that
Heidi added, or at least she didn't link to it when she committed the
tests.