Tweak suggestXyceTagList to better detect installed binaries
In an earlier commit, I had hacked in code to detect "installed"
binaries (those that were in "bin" directories instead of "src"
directories, so that it would automagically insert "-library?installed" to
disable the library tests when run outside of a build directory.

This was wrong in several ways.  Notably, it only called a binary
"installed" if it came from "make install", that is, if it contained a
"lib/libxyce.so" file or libxyce.a file and an include/N_CIR_Xyce.h.
If it found the shared library it added "?shared" as well as
"-library?installed".

But our standard binary builds have bin directories and lib
directories, but don't install libxyce.  They still need
"-library?installed" and not "?shared".  It does now.

Mac OS X has ".dylib" instead, and so we need to detect that as well
as .so files.  I had already fixed that problem in the VERILOGPLUGIN
testscripts, but not here.