Make VERILOGPLUGIN tests check for presence of admsXml
These scripts looked for the Xyce "buildxyceplugin" script, and were
intended to exit with an informative error if it was not found.  I
fixed that in the previous commit, but I never actually had checked
that the admsXml program (required by buildxyceplugin) is in the
user's path.

Thus, buildxyceplugin would get run anyway and just fail.

This commit makes the two tests check for admsXml and report that as
the error instead of leaving it for later.