Increase abstol on three signals
Three of the signals in the rlc_series.cir needed fairly loose
tolerance (1e-6) to pass on all platforms up until today.  Attaway is
showing these same three signals failing today by a very small
amount (1.5% or thereabouts).

Increasing abstol to 2e-6 fixes it.

When this test fails for REAL, it fails by 70%, so this small abstol
change does not break the test for its intended purpose (to verify
that __XYCE__ is defined when building the verilog).

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/158