Update tests to for new default random op behavior
This is related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/178
"Change default behavior of random operators (AGAUSS, GAUSS, etc) to
return the mean when not doing sampling."

This commit excludes four tests that were designed for the old default
random operator behavior, where operators like AGAUSS or RAND would
return a random number even if not doing sampling.   Very similar tests
to these already exist, in the SAMPLING2 directory; the difference being
that the SAMPLING2 tests have sampling commands in them.

This commit also updates the PARSER/random test so that (1) it no longer
attempts to use the -hspice-ext random command line option (which no
longer exists) and (2) tests all the random operators, not just AGAUSS
and GAUSS.  So, it now also tests RAND, UNIF and AUNIF.