Add more testing of variable .param
Related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also this bugzilla issue:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

This commit adds 3 tests.
1. paramStep1.cir, which uses .param with .step
2. paramHomotopy.cir, which uses .param with loca.
3. paramDc1.cir, same as paramStep1.cir, but for DC.

For now I have to exclude #3 b/c I haven't (yet) figured out how to
make the test script accept that the .param is a valid sweep variable on
the .print line.  But when I check this circuit manually it does the
right thing.