Add sampling and embedded sampling tests w/.param
Related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

Also this bugzilla issue:
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1190
"Re-think whether .PARAM statements need to evaluate to a constant"

This commit adds several tests of .param with both kinds of sampling in
Xyce.  In other words, .sampling and .embeddedsampling.  These are all
copies of tests that were originally designed with .global_params, but
using .param instead.