add issue 191 tests, exclamation point in params
Related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/191
"Support the exclamation point character in parameter names,
in the expression library"

Exclamation points can appear in fully resolved parameter names if they
are parameters from a "Y" device.  One oversight in the new expression
library was to not allow exclamation points in parameter names.  This
was not caught last year because there were no regression tests that
depended on this working.

This commit adds two tests, one that uses the "lumped line" Y device
and the other which uses the RXN Y device.