Add shared plugin tests for Xyce#191
@erkeite has fixed an expression library issue that first turned up in
a customer use of Verilog-A plugins, namely, that it was not possible
to use parameters of those devices in expressions such as
{YFOO!FOO1:param}.  The issue was that the expression library choked
on the exclamation point in the device name, but only when the
identifier was a bare word, not when it was identifying a node name or
internal variable name used in an N() construct.

Eric has created a number of tests to make sure this doesn't get
borken again, but I'm adding a few expressions involving "rlc" device
parameters to the VERILOGPLUGIN tests to test the fix in the context
that originally led to the issue being addressed.

Assuming this runs correctly in nightly testing, I'll be considering
it part of my QA for that issue closure.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/191