Adding valgrind option to generate suppressions for errors
Adding --gen-suppressions=all will include the necessary suppressions
in the output, if there are errors that are not covered by the provided
suppressions file.  This is useful in the event that the error is being
generated by a TPL or something that is out of our control.  In that
case, the appropriate suppression can be grabbed from the output and
put in the Xyce.supp file.

Xyce#192