Add test .param with .sens and .data
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params available to the device package, so
globally scoped normal params will be equivalent to .global_param"

I overlooked .SENS as another use case in which global params
are used.  This commit adds a single test (patterned after an
identical one that used global params) that tests this out.

I also overlooked using it with .DATA.  This commit also adds a
test for this.  It is identical to the paramStep1 test, except that the
STEP sweep is set up via .DATA.