Add two very simple intrusive PCE tests.
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/167
"Make globally-scoped .params be equivalent to .global_param"

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/174
"Document all the PCE related methods in the guides for release 7.3"

There were only 2 tests in the test suite for the fully intrusive PCE
method.  This is the most experimental of the various PCE-related
methods in Xyce.    For the other PCE and UQ methods, I had some simple
tests that had analytical solutions.  This commit adapts one of those
simple analytical tests to the fully intrusive method.

While this was the right thing to do in any case, this came up in my
work for on issue#167, which was about making .param be equivalent to
.global_param in the top-level circuit context.  The code I implemented
to do this checks to see what analyses are present in the netlist, and
if any them can potentially modify parameters, the code then checks if
any of those parameters was specified as a .param and elevates it to
".global_param" status.   As this code is very specific to analysis
types, to fully verify the issue I needed at least one .PCE test case.

This commit adds that test case, with the paramUniformPCE.cir test,
where the uncertain parameter is a .param.

The other test that is added is simpleUniformPCE.cir which is identical
to the paramUniformPCE test, except that the uncertain parameter is a
.global_param.