Add intrusive PCE tests for random expr operators
Issue #84
"Connect AGAUSS, GAUSS and RAND from expression library to .SAMPLING,
.EMBEDDEDSAMPLING and .PCE"

Issue #174.
"Document all the PCE related methods in the guides for release 7.3"

I added tests for sampling and embedded sampling a while ago in support
of issue #84.  But to fully test it out I also needed to add some tests
of random expression operators(such as UNIF, AGAUSS, etc) for .PCE.

This commit also adds the issue84 tag to all the tests in the NISP2,
NI_REGRESSION2 and PCE2 subdirectories.