Update the nint test to not use Perl POSIX::round
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/139
"Add nint(), fmod() and db() (Hspice expression functions) to Xyce
expression library"

Back when I originally worked on this issue, I created regression tests
(as well as unit tests) for nint and fmod.  Both of them were based on
Perl shell scripts and did not have gold standards.

However, the Perl-based test that I created for nint failed on several
nightly platforms, and it turned out that the reason was that I was
relying on the Perl POSIX::round function.  That function is not
available in earlier versions of Perl, and apparently
some of our test machines use older Perl versions.

This commit replaces the perl-based test for nint with a gold-standard
based test.