Add a few more inteprolator tests
Xyce#80
"Implement spline function in the new expression library"

Xyce#107
"Expand expression table capability to handle more options, including
turning off breakpoints"

This commit adds some extra tests of various interpolators.  It also
tests a few extra use cases for table, which are covered (partially) by
Xyce#107.

The original tests in this directory all read in a table from a file,
and the applied whatever interpolator to them.  However, there are two
ways in the new expression library to handle this, for every
interpolator.  For example, both

akima("file.dat")
akimafile("file.dat")

both work.  Similar for all the other interpolators, including table.
So I wanted to make sure there was coverage of all of this.

Finally, for Xyce#107, I implemented a form of table called "fasttable",
which is identical to table, but has no breakpoints.   This test
"fasttable.cir" uses this option.  The data file being read in is a
sampled sinewave, so it shouldn't need breakpoints to produce a decent
answer, even with the PWL interpolation used by table.

Like with all the other options in this directory, adding the word
"file"S to the end of fasttable also works, and it is tested here as
well.

Note, the main point of "fasttable" was to produce a faster running PWL
table.  These tests don't really exercise the speed of this method, but
they at least provide some test coverage.