Create tests for .param use of random operators
Issue #200
"Using random operator like AGAUSS in a .param does not result in the
parameter being elevated to .global_param"

Issue #84
"Connect AGAUSS, GAUSS and RAND from expression library to .SAMPLING,
.EMBEDDEDSAMPLING and .PCE"

Issue #167
"Make globally-scoped .params be equivalent to .global_param"

Issue #174
"Document all the PCE related methods in the guides for release 7.3"

This issue came up while I was documenting PCE methods for the 7.3
release.   It pertains to all the issues I listed above.

When I fixed issue #167, which makes .param act the same as
.global_param (if they are in the global scope), I overlooked one use
case.  That use case is when a .param is setup as an expression using a
random operator such as AUNIF, *and* the netlist contains sampling
commands like .SAMPLING.  If that is the case, then these parameters
need to be elevated to global_param status.

I fixed this in the release-7.3 Xyce source code earlier this week, and
a few of the examples that I added to the guides rely on this working.

The main thing that remained was for me to create a few tests that use
this capability. That is what this commit is about.  All the tests I am
adding in this commit are duplicates of .global_param tests, with the
only difference being that the .global_param statements are now for
.param.  So, these should be pretty safe and be just as robust as the
tests from which they are copied.