Test indexing bugfix for  using .sens w/.four
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/119
"using .sens with .four can result in an indexing mistake"

I originally added a test for this issue last week but that test bombed
everywhere, unfortunately.  I then realized that the test upon which it
was based was excluded from testing, probably because it also bombed
everywhere.  So, that test (CMOS_inverter_sens_four2.cir) got excluded
by me last week and this issue has been un-tested since then.

This commit adds a new test for the issue.  Actually, it just extends
an existing test, bsim3sensFourFD_complicated.cir to be just a bit more
complicated.  I have verified that this more complicated version of this
test, when run in Xyce 7.2, exhibits the bug.    And, I have also
manually verified that this new version does NOT exhibit the bug on the
7.3 branch.

Basically, in this new version I added a bunch of gratuitous extra
outputs to the various .FOUR statements.  They were requesting that the
same signal, V(2), be evaluated using the .FOUR command.  So, all the
various V(2) outputs in the bsim3sensFourFD_complicated.cir.four file
should match each other.  In Xyce 7.3, they  match, while in Xyce 7.2
they do not, because of the indexing mistake that resulted in some of
the outputs being mislabeled.