Tests for naked expressions in .param
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/195
"Make (more) expressions that don't have any spaces in them (such as
agauss by itself) not need curly braces or quotes"

Small handful of tests for issue 195.  This tests out if expressions on
.param and .global_param lines are able to be parsed correctly even if
not surrounded by curly braces or quotes.  This is possible to do on
these lines b/c .param and .global_param have a clearly defined
delimeter already, which is the equals sign.