add tests of comparison operators and ternarys
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/195
"Make (more) expressions that don't have any spaces in them (such as
agauss by itself) not need curly braces or quotes"

This commit expands two of my new parser tests (nakedIf.cir and
nakedTernary.cir) so that they do more comprehensive testing of
all the possible comparison operators.  Creating these tests revealed a
bug that I fixed earlier this afternoon.

This commit also adds the tests table3.cir and table4.cir to the
ABM_SPLINES directory.  This was inspired by some of my documentation
work.   In the expression library, I made it so that either "tablefile"
or "table" coudld be used to read in tabular data from a file.  But in
the documentation I mistakenly said that you could only use "table" to
read in data from a file if it was explicitly inside of an expression
via curly braces.  That wasn't true.

Either "table" or "tablefile" can be used to read data files, and that
will work even without curly braces.   The table3.cir and table4.cir
netlists test this use case.