Test for model name problem
These tests are to certify that this issue is fixed:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/206

I while ago I modified the Xyce parser so that model binning was turned
on by default.  That was tracked with this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/144

In the code that I modified, I introduced a logic error.  That logic
error would cause problems when a device instance line referred to a
node name that matched a model name.  It would stop parsing too early,
if that field was intended to be a node name.

I discovered this while working with a recent (important) customer
circuit.

This issue potentially could happen for nodes named after model
names, and also nodes named after model-binning prefix strings.

So, this commit introduces several tests to cover all of those use
cases.  The first one (issue206.cir) has a resistor with a .model, and
the resistor is attached to a node "n", that matches a MOSFET model
name.    The parser should be smart enough to know that a resistor would
never have a model named "n", and also smart enough to know that "n"
appears too early in the line.  My fix makes sure that is the case.

The other two tests are with model binning.  In both cases, the MOSFET
that is using a binned model is attached to a node that could be
problematic.  In the first test (issue206binning1.cir), the node name
matches the model name prefix (nch).  In the second test
(issue206binning2.cir), the node name matches a fully resolved mosfet
model name (nch.1).

All of these pass with the recent code changes.