Addition test for issue #84
Issue #84
"Connect AGAUSS, GAUSS and RAND from expression library to .SAMPLING,
.EMBEDDEDSAMPLING and .PCE"

While doing QA on this bug, it was pointed out to me that there was use
case that my tests had not covered.  That use case was one where a
device parameter depended on an expression containing more than one
random operator.  Earlier in this project, that use case did not work.
It does now, but I had not created a test that proved it.

This test should do it.  It is an NISP test, in which  two equivalent
versions of the same circuit are run and the ouptut from both is
compared using "diff".  As this is using the NISP algorithm, the
calculation is relatively small and completely deterministic.  So, the
"diff" test should be safe to do.

In one version of the circuit, the two random operators are applied to
different .global_params.  These global_params are then  used in the
expression for another global_param, which is then applied to a resistor
value.  This netlist is the "gold" standard.

In the other version, there are no .params or .global_params in the
netlist.  The fully expanded expression, with both random operators, is
100% inline in the resistor instance line.  This is the case that
would have broken before, but works now.