Add transient tests for bug 1145 SON
These tests are for this issue:

https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1145

".GLOBAL_PARAM does not work in .IC statements"

I recently fixed the source code so that global_params can be used on
.IC statements.  This appears to work for most analysis statements.  I
mostly focused on transient for my work so far, and that works fine.

These test circuits were attached to the bug (and a duplicate bug).  The
tranRC.cir circuit is a really simple transient RC circuit, that is
initialized with .IC, and the value used by the .IC statement comes from
a global parameter.  This is roughly the simplest use case one could
possibly come up with for this bug.

The other test case, lotka_volterra.cir, is a circuit that I was working
on last year and is inspired by the Wikipedia description of the
Lotka-Volterra equations, also known as predator-prey equations:

https://en.wikipedia.org/wiki/Lotka%E2%80%93Volterra_equations

In the wikipedia entry, there are examples with a bunch of different
initial conditions.  I wanted to run all of those initial conditions in
one simulation in Xyce. So,  I initialized the problem with .IC and
then swept a global_parameter over the values on the wikipedia page.
This didn't work before but it works now.