Remove parallel tag for the tranRc.cir test
Related to this issue:
https://joseki-srn.sandia.gov/bugzilla/show_bug.cgi?id=1145
".GLOBAL_PARAM does not work in .IC statements"

This test is failing in parallel.  I am pretty sure I know
why it is failing but don't have time to fix it today.

It passes if I change the .global_param to a .param, but
that of course removes the whole point of the test.  It
also passes if I leave the .global_param in place, but then
add the following line to it:

.step ic1 1 1 1

Adding this .step line doesn't change the functionality of the test, as
the line just sets "ic1" to have the same value it would have had anyway.
 But, that line forces the "setParam" function to be called in the device
package, and that function call includes a parallel reduction.  So, that
makes the error go away.

But, it should not be necessary to do this.  I think I know how to
fix this, but don't have time to do that today.  So, removing from
parallel testing.