Fix XyceDataFile.py for python3 and binary rawfiles
I just tried to use "convertToPrn2.py" to convert Xyce and Spice
binary rawfiles to prn, and found that Python3 puked on some input
strings because it was trying to treat them as UTF-8 encoded data.

This boils down to some python3 vs. python2 relics in XyceDataFile.py.

Python2, while we're still stuck with it on RHEL7, is already over a year
past its official end of life and is disappearing from most reasonably
responsive operating systems.

Fortunately, convertToPrn2.py and XyceDataFile.py are not heavily used
in our test suite, and never for binary raw files.  This commit should
have no effect on nightly testing, but I will revert it if it breaks
anything.

It *does*, however, make this thing work correctly on binary raw files
with Python3.