Revert "Fix XyceDataFile.py for python3 and binary rawfiles"
This reverts commit 1feda32defce7d75b8668078b5a3bd14a486f207.

The fix was correct for Python3 but breaks 8 tests on RHEL7 because
we're still stuck using Python2 there.

Python3 is installed on our RHEL7 machines, but we are still stuck
using python2 because of the mixed signal stuff.

When that is fixed, we can revert this reversion and get with the
program.  Python 2 has been past its end of life for over a year, and
it is high time that all of our python scripts in Xyce work with
python 3.