Remove one of the error messages from this script
This test has been failing since yesterday, when I pushed a bunch of
parameter and/or expression code changes.

This test is an error exit test.  The netlist has a bunch of errors in
it, and the script checks to see if Xyce emits a prescribed list of
error messages.  With my recent changes, the very last error message
doesn't get emitted.  But the circuit still causes Xyce to exit and all
the other errors are emitted.

It looks as though the final error message simply doesn't get triggered
any more.  I think I know why, but for now I am simply modifying the
test, to make nightly testing stop complaining.   In the meantime, I'll
fix the code so that this error message will be triggered again.