Test MEASFAIL=<0|1> for Successful Measures
This commit adds a regression test for the use of .OPTIONS MEASURE
MEASFAIL=<0|1> for successful measures.  The test covers both the
measure types that have their own version of the printMeasureResult()
function, as well as ones that use the version from the Base class.
Additional work is needed on testing successful FOUR measures though.

Issue #221
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/221