Improve Test of MEASFAIL=<0|1> for Successful Measures
This commit improves the regression test for the use of .OPTIONS
MEASURE MEASFAIL=<0|1> for successful measures.  The FIND-AT measure
type is now tested.  In addition, the DERIV-AT, DERIV-WHEN, FIND-WHEN,
WHEN and FIND-AT measure types are now tested for TRAN_CONT measure
mode.  The only missing subtest is for the FOUR measure type.

Issue #221
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/221