Restore this test to include last failure message
This is related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/225
"Fix the MEASURE/bad_dot_measure_line2 test"

I just fixed the Xyce source code for this issue, so it should be safe
to restore this part of this test.  Due to some of my refactoring, one
of the error messages in this test was no longer being reported from
Xyce.  That was a mistake, and I've fixed it now.