Turn off 8 processor testing of lotka_volterra
Ever since it was added, the lotka_volterra test has been failing with
a timestep too small on 8 processors.  This is not fixed simply by
running with KLU as the solver.

Since it has been failing consistently for some weeks now, and is not
a high priority thing to fix, there is no benefit to having it fail
every night and email some random subset of the team about it; Jenkins
emails anyone who has committed since the last passing run of a job,
and having a single test fail once every day means that every single
person who committed that day gets emailed about it.

When the time comes to address this failure, it should be done
manually by whomever works on it, and not re-enabled for nightly
testing until it passes on 8 processors on at least one system.