Update error message tests to include {}
I recently changed the netlist parser so that it no longer strips {}
from expression strings before handing off to the expression library.
This seemed like a pointless thing to do, since the library can handle
curly braces just fine.

But, doing so caused 7 tests to break, because the error messages were
no longer precisely the same.  This updates the scripts for those 7
tests, so that they pass again.