Improve Tests for Precision in .MEASURE Output
This commit improves the regression tests for the PRECISION
qualifier and for .OPTIONS MEASURE MEASDGT=<value>.  The test
MEASURE/MeasdgtOption now covers all of the relevant measure types,
including TRAN_CONT ones.  It tests the precision in both the
<netlistName>.mt0 file and (now) stdout.  The MEASURE/PrecisionTest
covers the same measure types with the PRECISION qualifier given on
each .MEASURE line.  The default precision, in both the stdout and
the <netlistName>.mt0, is now tested by MEASURE/ErrorMessageTest.
Previously, the precision (and number format) was not tested in
stdout by MEASURE/ErrorMessageTest.

Issue #226
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/226