Test -o Output for .MEASURE, .FOUR and .FFT
This tests the use of the -o file name as the "base filename"
for the output of .MEASURE, .FOUR and .FFT lines.  It covers
both TRAN and TRAN_CONT measure modes, for the case where the
result for the continuous measure is sent to its own
<dashoFilename>_<measurename>.mtX file. This commit covers
both the .STEP and non-step cases.  So, it also verifies that the
results file is named <dashoFilename>.res file for the .STEP case.

The contents (and names) of the files for the .PRINT, .MEASURE and
.FFT output are verified, as is the results filename and content.
The contents of the .FOUR output files are not yet verified, but
the correctness of those file names(<dashoFilename>.fourX) is
verified.

Issue #222
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/222