Add I() and P() Operators to Test for Issue 229
Issue 229 was found in the context of V() operators.  However, that
problem also impacted I() and P() operators.  So, this commit adds
the use of I() and P() in expressions to the .PRINT TRAN line in the
existing test ISSUE_229/GlobalInExpressionBug.  This is mostly
because it was trivial to do so, while providing some additional
bug coverage.

Issue #229
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/229