Update Tests for -o Command Line Option Output
This commit updates the tests for the -o command line option.  The
comments in each .sh file provide more details.  However, the
general set of test requirements is that:

  1) the .PRINT outputs should go to the file(s) specified by -o, and
     not to the default file name or the file specified with FILE=.
     They should be FORMAT=STD files with space as the delimiter,
     with the correct default extensions (e.g., dashoFile.prn or
     dashoFile.SENS.prn for .PRINT TRAN or .PRINT SENS output).

  2) Print line concatenation should still work.

  3) the .sh files typically run their netlists with -delim COMMA
     to verify that command line option is ignored when -o is
     also specified.

The use of -r and -o on the same command line is tested.  Poorly
chosen -o <dashoFilename> that attempt to overwrite the <netlistFile>
are also tested for the .PRINT TRAN and .MEASURE TRAN cases.
Finally, an attempt to set <dashoFilename> is verified to produce
a netlist parsing error.

Issue #222
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/222