test ckts for use of scale and dependent params
This is related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/239

"Make ".options parser scale=#" play nice with dependent parameters"

For all of these tests, they are copies of previously existing tests for
"options parser scale".  The only difference is that in the new tests,
I've modified every model card to that there is at least one parameter
that is using agauss.  This was the use case the triggered the bug; if
"scale" was being used, and the model also included at least one
"agauss" parameter, then the model would get set up more than once, and
that would cause "scale" to be applied more than once, which was wrong.