Update XDM Tests for New -o Command Line Option Output
This commit updates the XDM tests to reflect the recent changes
to the -o command line option output.  The <dashoFilename> is
now the "base name" for the output files.  So, various Xyce
invocations (that use -o) in XdmCommon.pm or some .sh files
needed to be changed to reflect that.  Issue 222 provides
more details.

Issue #222
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/222