Tests for modifying scaled parameters via sampling
and also .STEP.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/239
"Make ".options parser scale=#" play nice with dependent parameters"

This commits adds several tests.   All of them test different variations of
the same idea, which is for Xyce to correctly run an analysis (such as
.STEP and/or .SAMPLING) which modifies scaled parameters such as a
MOSFET L,W.   If these params are being updated, then they need to have
the scale factors applied at each update.

All 3 of these tests have the same structure.  In each case, there is a
scaled version and an unscaled version, and they are compared.

The MOSFET device models covered in this commit include the BSIM3 and
BSIM4.

I will also note that the tests over in the
Xyce_Regression/Netlists/Certification_Tests/BUG_1041_SON directory are
also applicable to this issue.