Adds a new test for hierarchical circuits that are missing '.ENDS'
Xyce has been testing that flattened circuits exit cleanly with an
error if a subcircuit is missing the '.ENDS' termination.  However,
hierarchical circuits were not being tested.  This caused an issue,
an infinite loop, in the parsing of a very large ASIC. A check in
the code has been added to throw an error if this happens.  This
new test makes sure the added check works.

https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/245