Update Regression Tests for -o Command Line Option
This commit updates various tests in Output/Dasho to test that
the "default print extension" is removed if it is the suffix of
the requested -o filename.  So, for example, this command line:

  Xyce -o acOutput.FD.prn netlist.cir

will use acOutput as the dashoFilename.  Similar changes were made
to the tests for .ES, .HB, .NOISE, .PCE and .LIN analyses.

Issue #222
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/222