Update the OP output gold standard :: -> :
This is (vaguely) related to this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/223
"Add a new subcircuit separator, "." instead of ":" as part of
-hspice-ext"

I recently found and fixed a minor issue in the parser, in which the
subcircuit separator was being applied 2x for .MODEL statements declared
inside of subcircuits.  This 2x application resulted in model names like
subckt1::modelName, which should have instead been subckt1:modelName.

The only test that broke as a result of my fix was the sand25OpAlone
test.  This commit fixes the gold standard for that test to just have
single separators