Add separator tests for .MODELs inside subckts
This is in support of this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/223
"Add a new subcircuit separator, "." instead of ":" as part of
-hspice-ext"

These are very similar tests to the other separator tests I added
previously.    The focus of these tests are the full name resolution
(with the new period separator) of model parameters which are defined
inside of subcircuit definitions.

There are two possible use cases, and one is tested by the separator5
test, and the other is tested by the separator6 test.

1. Subcircuit .MODEL cards which are 100% constant, and not dependent on
   subcircuit arguments or any other expressions.  (separator5 test)
   In this case, Xyce will only create a single model, which is used
   by all the subcircuit instances.  In this case, the subcircuit model
   name prefix is from the subcircuit definition.

2. Subcircuit .MODEL cards which include at least one parameter that has
   a subcircuit parameter and/or expression dependence.(separator6 test)
   In this case, Xyce will only create a unique model for each
   subcircuit instance.  In this case, the subcircuit model name
   prefix is from the subcircuit instance name.