Add more tests for the optional period separator
This is in support of this issue:
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/223
"Add a new subcircuit separator, "." instead of ":" as part of
-hspice-ext"

Pete suggested that I look at the various tests in the INVALID_CHARS
directory.  There were several that were relevant to this issue.
When running with the new hspice separator option, the rules for valid
characters translate pretty directly.

So, whereas before a colon was NOT allowed in a device name, but a
period was allowed, it is the other way around when using this
separator.  Similarly, colons were allowed in node names before, even
though it was a very bad idea, and with this separator, periods are
allowed in node names, even though it is a very bad idea.

The new tests added all reflect these changes, and are forced to run
using the command line option for hspice separators.