Modify .sh File for Output/ES/es-regr-pce-tran-prn
This commit modifies the .sh file for Output/ES/es-regr-pce-tran-prn
to (essentially) run the test a second time, if the first time fails
comparison of the es-regr-pce-tran-prn.ES.prn file.  This is done
because sometimes the first random sample is somehow picking a "bad"
configuration" that outputs a result too far away from values in
the gold standard.  This approach is used by other tests in
EMBEDDED.

Issue 247
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/247