Improve Tests for Analysis vs. .PRINT Checking
This commit re-orders the .PRINT TRANADJOINT and .PRINT TRAN
statements in this test's netlist.  This tests that their
ordering does not matter to the checking of the analysis type
vs. the .PRINT line type.

It also modifies Certification_Tests/BUG_80/op_test to test
that multiple incorrect .PRINT lines are flagged, even if
they come after a correct .PRINT line.

SON Bug 1029
https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1029