Test for Invalid Combo of Start-End Sim Times
This commit adds an errorexit test for the case of the end
simulation time being before the start simulation time.  This
error can occur when manually converting an HSPICE .TRAN
statement to a Xyce .TRAN statement.  The relevant error
message already existed in the source code, but was not tested.