Refactor Various Perl-Based Comparators to use strict
This commit adds use strict to various Perl-based comparators that
are leveraged by many .sh files in the regression test suite. It is
an offshoot of the work on improving the regression tests for .FOUR
lines.  Adding use strict did uncover one coding error (in the debug
output) for ACComparator.pl.

Issue #248
https://gitlab-ex.sandia.gov/Xyce/Xyce/-/issues/248